pyproject.toml
[project]
name = "myhelloagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-search-results>=2.4.2",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
]
